chore(discord): shift entity name modal to sponsor layout
